pyproject.toml
[tool.poetry]
name = "book"
version = "0.1.0"
description = ""
authors = ["Arsen <arsen8577711@yandex.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
pytest = "^7.3.1"
djangorestframework = "^3.14.0"
djangorestframework-simplejwt = "^5.2.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
